WIP: Refactored approach to shaking items when quick stack adds items to the current chest menu. Should be safer and simpler now.
