+++ ConvenientInventory/QuickStack/QuickStackLogic.cs
@@ -2,7 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using ConvenientInventory.AutoOrganize;
 using ConvenientInventory.TypedChests;
 using Microsoft.Xna.Framework;
 using StardewModdingAPI;
@@ -17,203 +16,6 @@ namespace ConvenientInventory.QuickStack
 {
     public static class QuickStackLogic
     {
-        public static bool IsStackingIntoItemGrabMenu { get; private set; }
-
-        public static bool StackInItemGrabMenu(ItemGrabMenu itemGrabMenu)
-        {
-            try
-            {
-                IsStackingIntoItemGrabMenu = true;
-                return StackInItemGrabMenuCore(itemGrabMenu);
-            }
-            finally
-            {
-                IsStackingIntoItemGrabMenu = false;
-            }
-        }
-
-        private static bool StackInItemGrabMenuCore(ItemGrabMenu itemGrabMenu)
-        {
-            Chest chest = itemGrabMenu.context switch
-            {
-                Chest c => c,
-                JunimoHut jh => jh.GetOutputChest(),
-                _ => null,
-            };
-
-            if (chest is null)
-            {
-                Game1.playSound("cancel");
-                ModEntry.Instance.Monitor.Log(
-                    $"Cannot quick stack into ItemGrabMenu context of type '{itemGrabMenu.context?.GetType().Name ?? "null"}'.",
-                    LogLevel.Debug);
-                return false;
-            }
-
-            if (!ShouldQuickStackInto(chest, out ChestType chestType, false))
-            {
-                Game1.playSound("cancel");
-                return false;
-            }
-            
-            //TODO: Add config option to allow manual quick stack into chest to bypass a "disabled" quick stack toggle state, true by default.
-            if (QuickStackToggleChestLogic.GetQuickStackToggleChestStateFromModData(chest) == QuickStackToggleChestState.Disabled)
-            {
-                //TODO: Shake Toggle Quick Stack button, if visible.
-                Game1.playSound("cancel");
-                return false;
-            }
-
-            TypedChest typedChest = new(
-                chest, itemGrabMenu.context is JunimoHut ? ChestType.JunimoHut : chestType, chest.Location, null);
-
-            bool movedAtLeastOneTotal = false;
-            Farmer who = Game1.player;
-            Inventory playerInventory = who.Items;
-
-            QuickStackSummary quickStackSummary = new();
-            List<Item> overflowItems = new();
-
-            // Fill chest stacks with player inventory items
-            foreach (Item chestItem in chest.GetItemsForPlayer(who.UniqueMultiplayerID))
-            {
-                if (chestItem is null)
-                {
-                    continue;
-                }
-
-                foreach (Item playerItem in playerInventory)
-                {
-                    if (playerItem is null)
-                    {
-                        continue;
-                    }
-
-                    if (ModEntry.Config.FavoriteItems.IsEnabled && ConvenientInventory.FavoriteItemSlots[playerInventory.IndexOf(playerItem)])
-                    {
-                        // Skip favorited items
-                        continue;
-                    }
-
-                    if (!playerItem.canStackWith(chestItem))
-                    {
-                        if (ModEntry.Config.QuickStack.OverflowItems
-                            && ModEntry.Config.QuickStack.IgnoreItemQuality
-                            && CanStackWithIgnoreQuality(playerItem, chestItem))
-                        {
-                            overflowItems.Add(playerItem.getOne());
-                        }
-
-                        continue;
-                    }
-
-                    int beforeStack = playerItem.Stack;
-                    playerItem.Stack = chestItem.addToStack(playerItem);
-                    bool movedAtLeastOne = beforeStack != playerItem.Stack;
-
-                    movedAtLeastOneTotal |= movedAtLeastOne;
-
-                    if (movedAtLeastOne)
-                    {
-                        itemGrabMenu.ItemsToGrabMenu.ShakeItem(chestItem);
-
-                        ClickableComponent inventoryComponent = itemGrabMenu.inventory.inventory[playerInventory.IndexOf(playerItem)];
-                        itemGrabMenu._transferredItemSprites.Add(
-                            new ItemGrabMenu.TransferredItemSprite(playerItem.getOne(), inventoryComponent.bounds.X, inventoryComponent.bounds.Y));
-
-                        if (playerItem.Stack == 0)
-                        {
-                            who.removeItemFromInventory(playerItem);
-                        }
-
-                        quickStackSummary.AddToSummary(typedChest, playerItem.Name, playerItem.Stack, beforeStack);
-                    }
-
-                    if (chestItem.Stack == chestItem.maximumStackSize())
-                    {
-                        if (ModEntry.Config.QuickStack.OverflowItems)
-                        {
-                            overflowItems.Add(chestItem.getOne());
-                        }
-
-                        itemGrabMenu.inventory.ShakeItem(playerItem);
-                        break;
-                    }
-                }
-            }
-
-            // Add overflow stacks to chest when applicable
-            IInventory chestItems = chest.GetItemsForPlayer(who.UniqueMultiplayerID);
-            if (ModEntry.Config.QuickStack.OverflowItems && chestItems.Count < chest.GetActualCapacity())
-            {
-                foreach (Item overflowItem in overflowItems)
-                {
-                    if (overflowItem is null)
-                    {
-                        continue;
-                    }
-
-                    foreach (Item playerItem in playerInventory)
-                    {
-                        if (playerItem is null)
-                        {
-                            continue;
-                        }
-
-                        if (ModEntry.Config.FavoriteItems.IsEnabled && ConvenientInventory.FavoriteItemSlots[playerInventory.IndexOf(playerItem)])
-                        {
-                            // Skip favorited items
-                            continue;
-                        }
-
-                        if (!playerItem.canStackWith(overflowItem))
-                        {
-                            // Skip overflow item if it doesn't stack with player item.
-                            continue;
-                        }
-
-                        int beforeStack = playerItem.Stack;
-                        Item leftoverItem = chest.addItem(playerItem);
-                        bool movedAtLeastOne = leftoverItem is null || beforeStack != leftoverItem.Stack;
-
-                        movedAtLeastOneTotal |= movedAtLeastOne;
-
-                        if (movedAtLeastOne)
-                        {
-                            ClickableComponent inventoryComponent = itemGrabMenu.inventory.inventory[playerInventory.IndexOf(playerItem)];
-                            itemGrabMenu._transferredItemSprites.Add(
-                                    new ItemGrabMenu.TransferredItemSprite(playerItem.getOne(), inventoryComponent.bounds.X, inventoryComponent.bounds.Y));
-
-                            quickStackSummary.AddToSummary(typedChest, playerItem.Name, leftoverItem?.Stack ?? 0, beforeStack);
-                        }
-
-                        if (leftoverItem is null)
-                        {
-                            who.removeItemFromInventory(playerItem);
-                        }
-                        else
-                        {
-                            itemGrabMenu.inventory.ShakeItem(playerItem);
-                        }
-                    }
-                }
-            }
-
-            Game1.playSound(movedAtLeastOneTotal ? "Ship" : "cancel");
-
-            if (movedAtLeastOneTotal)
-            {
-                if (ModEntry.Config.AutoOrganizeChest.IsEnabled)
-                {
-                    AutoOrganizeLogic.TryOrganizeChestOnFillOutStacks(itemGrabMenu, chest);
-                }
-
-                ModEntry.Instance.Monitor.Log(quickStackSummary.GetSummaryMessage(), LogLevel.Trace);
-            }
-
-            return movedAtLeastOneTotal;
-        }
-
         public static bool StackToNearbyChests(string rangeStr, InventoryPage inventoryPage = null)
         {
             bool movedAtLeastOneTotal = false;
@@ -499,6 +301,48 @@ namespace ConvenientInventory.QuickStack
                 return false;
             }
             else if (!ModEntry.Config.QuickStack.IntoMiniShippingBins && chestType == ChestType.MiniShippingBin)
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        // Taken from `Item.canStackWith`, removing quality check.
+        public static bool CanStackWithIgnoreQuality(Item item, ISalable other)
+        {
+            if (other is not Item otherItem || other.GetType() != item.GetType())
+            {
+                return false;
+            }
+
+            if (item is ColoredObject coloredObj)
+            {
+                if (other is ColoredObject otherColoredObj && !coloredObj.color.Value.Equals(otherColoredObj.color.Value))
+                {
+                    return false;
+                }
+            }
+
+            if (item.maximumStackSize() <= 1 || other.maximumStackSize() <= 1)
+            {
+                return false;
+            }
+
+            if (item is StardewValley.Object obj)
+            {
+                if (other is StardewValley.Object otherObj && otherObj.orderData.Value != obj.orderData.Value)
+                {
+                    return false;
+                }
+            }
+
+            if (item.QualifiedItemId != otherItem.QualifiedItemId)
+            {
+                return false;
+            }
+
+            if (!item.Name.Equals(other.Name))
             {
                 return false;
             }
@@ -1058,48 +902,6 @@ namespace ConvenientInventory.QuickStack
         private static bool IsPositionWithinRange(Vector2 origin, Vector2 target, int range)
         {
             return Math.Abs(origin.X - target.X) <= range * Game1.tileSize && Math.Abs(origin.Y - target.Y) <= range * Game1.tileSize;
-        }
-
-        // Taken from `Item.canStackWith`, removing quality check.
-        private static bool CanStackWithIgnoreQuality(Item item, ISalable other)
-        {
-            if (other is not Item otherItem || other.GetType() != item.GetType())
-            {
-                return false;
-            }
-
-            if (item is ColoredObject coloredObj)
-            {
-                if (other is ColoredObject otherColoredObj && !coloredObj.color.Value.Equals(otherColoredObj.color.Value))
-                {
-                    return false;
-                }
-            }
-
-            if (item.maximumStackSize() <= 1 || other.maximumStackSize() <= 1)
-            {
-                return false;
-            }
-
-            if (item is StardewValley.Object obj)
-            {
-                if (other is StardewValley.Object otherObj && otherObj.orderData.Value != obj.orderData.Value)
-                {
-                    return false;
-                }
-            }
-
-            if (item.QualifiedItemId != otherItem.QualifiedItemId)
-            {
-                return false;
-            }
-
-            if (!item.Name.Equals(other.Name))
-            {
-                return false;
-            }
-
-            return true;
         }
     }
 }
